changed parameters in API-DTOs removed; package exception moved
