+++ wls-eai-service/src/main/java/de/muenchen/oss/wahllokalsystem/eaiservice/rest/wahlvorschlag/dto/ReferendumvorlagenDTO.java
@@ -4,8 +4,6 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.util.Set;
 
-public record ReferendumvorlagenDTO(@NotNull String wahlbezirkID,
-                                    @NotNull String wahlID,
-                                    @NotNull String stimmzettelgebietID,
+public record ReferendumvorlagenDTO(@NotNull String stimmzettelgebietID,
                                     @NotNull @Size(min = 1) Set<ReferendumvorlageDTO> referendumvorlagen) {
 }
