+++ wls-eai-service/src/main/java/de/muenchen/oss/wahllokalsystem/eaiservice/rest/wahlvorschlag/dto/WahlvorschlaegeListeDTO.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.util.Set;
 
-public record WahlvorschlaegeListeDTO(@NotNull String wahltag, @NotNull String wahlID,
+public record WahlvorschlaegeListeDTO(@NotNull String wahlID,
                                       @NotNull @Size(min = 1) Set<WahlvorschlaegeDTO> wahlvorschlaegeliste) {
 }
