+++ wls-eai-service/src/test/java/de/muenchen/oss/wahllokalsystem/eaiservice/service/wahlvorschlag/WahlvorschlagMapperTest.java
@@ -43,12 +43,12 @@ class WahlvorschlagMapperTest {
             val result = unitUnderTest.toDTO(entityToMap);
 
             val expectedKandidaten = Set.of(
-                    new KandidatDTO(kandidat1.getId().toString(), kandidat1.getName(), kandidat1.getListenposition(), kandidat1.isDirektkandidat(),
-                            kandidat1.getTabellenSpalteInNiederschrift(), kandidat1.isEinzelbewerber()),
-                    new KandidatDTO(kandidat2.getId().toString(), kandidat2.getName(), kandidat2.getListenposition(), kandidat2.isDirektkandidat(),
-                            kandidat2.getTabellenSpalteInNiederschrift(), kandidat2.isEinzelbewerber()));
+                new KandidatDTO(kandidat1.getId().toString(), kandidat1.getName(), kandidat1.getListenposition(), kandidat1.isDirektkandidat(),
+                    kandidat1.getTabellenSpalteInNiederschrift(), kandidat1.isEinzelbewerber()),
+                new KandidatDTO(kandidat2.getId().toString(), kandidat2.getName(), kandidat2.getListenposition(), kandidat2.isDirektkandidat(),
+                    kandidat2.getTabellenSpalteInNiederschrift(), kandidat2.isEinzelbewerber()));
             val expectedResult = new WahlvorschlagDTO(entityToMap.getId().toString(), entityToMap.getOrdnungszahl(), entityToMap.getKurzname(),
-                    entityToMap.isErhaeltStimmen(), expectedKandidaten);
+                entityToMap.isErhaeltStimmen(), expectedKandidaten);
 
             Assertions.assertThat(result).isEqualTo(expectedResult);
         }
@@ -76,23 +76,23 @@ class WahlvorschlagMapperTest {
             val result = unitUnderTest.toDTO(wahlvorschlaege);
 
             val expectedKandidaten1 = Set.of(
-                    new KandidatDTO(kandidat1.getId().toString(), kandidat1.getName(), kandidat1.getListenposition(), kandidat1.isDirektkandidat(),
-                            kandidat1.getTabellenSpalteInNiederschrift(), kandidat1.isEinzelbewerber()),
-                    new KandidatDTO(kandidat2.getId().toString(), kandidat2.getName(), kandidat2.getListenposition(), kandidat2.isDirektkandidat(),
-                            kandidat2.getTabellenSpalteInNiederschrift(), kandidat2.isEinzelbewerber()));
+                new KandidatDTO(kandidat1.getId().toString(), kandidat1.getName(), kandidat1.getListenposition(), kandidat1.isDirektkandidat(),
+                    kandidat1.getTabellenSpalteInNiederschrift(), kandidat1.isEinzelbewerber()),
+                new KandidatDTO(kandidat2.getId().toString(), kandidat2.getName(), kandidat2.getListenposition(), kandidat2.isDirektkandidat(),
+                    kandidat2.getTabellenSpalteInNiederschrift(), kandidat2.isEinzelbewerber()));
             val expectedWahlvorschlag1 = new WahlvorschlagDTO(wahlvorschlag1.getId().toString(), wahlvorschlag1.getOrdnungszahl(), wahlvorschlag1.getKurzname(),
-                    wahlvorschlag1.isErhaeltStimmen(), expectedKandidaten1);
+                wahlvorschlag1.isErhaeltStimmen(), expectedKandidaten1);
             val expectedKandidaten2 = Set.of(
-                    new KandidatDTO(kandidat3.getId().toString(), kandidat3.getName(), kandidat3.getListenposition(), kandidat3.isDirektkandidat(),
-                            kandidat3.getTabellenSpalteInNiederschrift(), kandidat3.isEinzelbewerber()),
-                    new KandidatDTO(kandidat4.getId().toString(), kandidat4.getName(), kandidat4.getListenposition(), kandidat4.isDirektkandidat(),
-                            kandidat4.getTabellenSpalteInNiederschrift(), kandidat4.isEinzelbewerber()));
+                new KandidatDTO(kandidat3.getId().toString(), kandidat3.getName(), kandidat3.getListenposition(), kandidat3.isDirektkandidat(),
+                    kandidat3.getTabellenSpalteInNiederschrift(), kandidat3.isEinzelbewerber()),
+                new KandidatDTO(kandidat4.getId().toString(), kandidat4.getName(), kandidat4.getListenposition(), kandidat4.isDirektkandidat(),
+                    kandidat4.getTabellenSpalteInNiederschrift(), kandidat4.isEinzelbewerber()));
             val expectedWahlvorschlag2 = new WahlvorschlagDTO(wahlvorschlag2.getId().toString(), wahlvorschlag2.getOrdnungszahl(), wahlvorschlag2.getKurzname(),
-                    wahlvorschlag2.isErhaeltStimmen(), expectedKandidaten2);
+                wahlvorschlag2.isErhaeltStimmen(), expectedKandidaten2);
 
             val expectedResult = new WahlvorschlaegeDTO(wahlvorschlaege.getWahlbezirkID(), wahlvorschlaege.getWahlID(),
-                    wahlvorschlaege.getStimmzettelgebietID(),
-                    Set.of(expectedWahlvorschlag1, expectedWahlvorschlag2));
+                wahlvorschlaege.getStimmzettelgebietID(),
+                Set.of(expectedWahlvorschlag1, expectedWahlvorschlag2));
             Assertions.assertThat(result).isEqualTo(expectedResult);
         }
 
@@ -120,25 +120,25 @@ class WahlvorschlagMapperTest {
             val result = unitUnderTest.toDTO(wahlvorschlaegeListe);
 
             val expectedKandidaten1 = Set.of(
-                    new KandidatDTO(kandidat1.getId().toString(), kandidat1.getName(), kandidat1.getListenposition(), kandidat1.isDirektkandidat(),
-                            kandidat1.getTabellenSpalteInNiederschrift(), kandidat1.isEinzelbewerber()),
-                    new KandidatDTO(kandidat2.getId().toString(), kandidat2.getName(), kandidat2.getListenposition(), kandidat2.isDirektkandidat(),
-                            kandidat2.getTabellenSpalteInNiederschrift(), kandidat2.isEinzelbewerber()));
+                new KandidatDTO(kandidat1.getId().toString(), kandidat1.getName(), kandidat1.getListenposition(), kandidat1.isDirektkandidat(),
+                    kandidat1.getTabellenSpalteInNiederschrift(), kandidat1.isEinzelbewerber()),
+                new KandidatDTO(kandidat2.getId().toString(), kandidat2.getName(), kandidat2.getListenposition(), kandidat2.isDirektkandidat(),
+                    kandidat2.getTabellenSpalteInNiederschrift(), kandidat2.isEinzelbewerber()));
             val expectedWahlvorschlag1 = new WahlvorschlagDTO(wahlvorschlag1.getId().toString(), wahlvorschlag1.getOrdnungszahl(), wahlvorschlag1.getKurzname(),
-                    wahlvorschlag1.isErhaeltStimmen(), expectedKandidaten1);
+                wahlvorschlag1.isErhaeltStimmen(), expectedKandidaten1);
             val expectedKandidaten2 = Set.of(
-                    new KandidatDTO(kandidat3.getId().toString(), kandidat3.getName(), kandidat3.getListenposition(), kandidat3.isDirektkandidat(),
-                            kandidat3.getTabellenSpalteInNiederschrift(), kandidat3.isEinzelbewerber()),
-                    new KandidatDTO(kandidat4.getId().toString(), kandidat4.getName(), kandidat4.getListenposition(), kandidat4.isDirektkandidat(),
-                            kandidat4.getTabellenSpalteInNiederschrift(), kandidat4.isEinzelbewerber()));
+                new KandidatDTO(kandidat3.getId().toString(), kandidat3.getName(), kandidat3.getListenposition(), kandidat3.isDirektkandidat(),
+                    kandidat3.getTabellenSpalteInNiederschrift(), kandidat3.isEinzelbewerber()),
+                new KandidatDTO(kandidat4.getId().toString(), kandidat4.getName(), kandidat4.getListenposition(), kandidat4.isDirektkandidat(),
+                    kandidat4.getTabellenSpalteInNiederschrift(), kandidat4.isEinzelbewerber()));
             val expectedWahlvorschlag2 = new WahlvorschlagDTO(wahlvorschlag2.getId().toString(), wahlvorschlag2.getOrdnungszahl(), wahlvorschlag2.getKurzname(),
-                    wahlvorschlag2.isErhaeltStimmen(), expectedKandidaten2);
+                wahlvorschlag2.isErhaeltStimmen(), expectedKandidaten2);
             val expectedWahlvorschlaege = new WahlvorschlaegeDTO(wahlvorschlaege.getWahlbezirkID(), wahlvorschlaege.getWahlID(),
-                    wahlvorschlaege.getStimmzettelgebietID(),
-                    Set.of(expectedWahlvorschlag1, expectedWahlvorschlag2));
+                wahlvorschlaege.getStimmzettelgebietID(),
+                Set.of(expectedWahlvorschlag1, expectedWahlvorschlag2));
 
-            val expectedResult = new WahlvorschlaegeListeDTO(wahlvorschlaegeListe.getWahltag().toString(), wahlvorschlaegeListe.getWahlID(),
-                    Set.of(expectedWahlvorschlaege));
+            val expectedResult = new WahlvorschlaegeListeDTO(wahlvorschlaegeListe.getWahlID(),
+                Set.of(expectedWahlvorschlaege));
 
             Assertions.assertThat(result).isEqualTo(expectedResult);
         }
@@ -171,10 +171,10 @@ class WahlvorschlagMapperTest {
             val result = unitUnderTest.toDTO(entityToMap);
 
             val expectedReferendumoptionen = Set.of(
-                    new ReferendumoptionDTO(referendumoption1.getId().toString(), referendumoption1.getName(), referendumoption1.getPosition()),
-                    new ReferendumoptionDTO(referendumoption2.getId().toString(), referendumoption2.getName(), referendumoption2.getPosition()));
+                new ReferendumoptionDTO(referendumoption1.getId().toString(), referendumoption1.getName(), referendumoption1.getPosition()),
+                new ReferendumoptionDTO(referendumoption2.getId().toString(), referendumoption2.getName(), referendumoption2.getPosition()));
             val expectedResult = new ReferendumvorlageDTO(entityToMap.getWahlvorschlagID(), entityToMap.getOrdnungszahl(), entityToMap.getKurzname(),
-                    entityToMap.getFrage(), expectedReferendumoptionen);
+                entityToMap.getFrage(), expectedReferendumoptionen);
 
             Assertions.assertThat(result).isEqualTo(expectedResult);
         }
@@ -202,20 +202,20 @@ class WahlvorschlagMapperTest {
             val result = unitUnderTest.toDTO(entityToMap);
 
             val expectedReferendumoptionen1 = Set.of(
-                    new ReferendumoptionDTO(referendumoption1.getId().toString(), referendumoption1.getName(), referendumoption1.getPosition()),
-                    new ReferendumoptionDTO(referendumoption2.getId().toString(), referendumoption2.getName(), referendumoption2.getPosition()));
+                new ReferendumoptionDTO(referendumoption1.getId().toString(), referendumoption1.getName(), referendumoption1.getPosition()),
+                new ReferendumoptionDTO(referendumoption2.getId().toString(), referendumoption2.getName(), referendumoption2.getPosition()));
             val expectedReferendumoptionen2 = Set.of(
-                    new ReferendumoptionDTO(referendumoption3.getId().toString(), referendumoption3.getName(), referendumoption3.getPosition()),
-                    new ReferendumoptionDTO(referendumoption4.getId().toString(), referendumoption4.getName(), referendumoption4.getPosition()));
+                new ReferendumoptionDTO(referendumoption3.getId().toString(), referendumoption3.getName(), referendumoption3.getPosition()),
+                new ReferendumoptionDTO(referendumoption4.getId().toString(), referendumoption4.getName(), referendumoption4.getPosition()));
             val expectedReferendumvorlage1 = new ReferendumvorlageDTO(referendumvorlage1.getWahlvorschlagID(), referendumvorlage1.getOrdnungszahl(),
-                    referendumvorlage1.getKurzname(),
-                    referendumvorlage1.getFrage(), expectedReferendumoptionen1);
+                referendumvorlage1.getKurzname(),
+                referendumvorlage1.getFrage(), expectedReferendumoptionen1);
             val expectedReferendumvorlage2 = new ReferendumvorlageDTO(referendumvorlage2.getWahlvorschlagID(), referendumvorlage2.getOrdnungszahl(),
-                    referendumvorlage2.getKurzname(),
-                    referendumvorlage2.getFrage(), expectedReferendumoptionen2);
+                referendumvorlage2.getKurzname(),
+                referendumvorlage2.getFrage(), expectedReferendumoptionen2);
 
-            val expectedResult = new ReferendumvorlagenDTO(entityToMap.getWahlbezirkID(), entityToMap.getWahlID(),
-                    entityToMap.getStimmzettelgebietID(), Set.of(expectedReferendumvorlage1, expectedReferendumvorlage2));
+            val expectedResult = new ReferendumvorlagenDTO(entityToMap.getStimmzettelgebietID(),
+                Set.of(expectedReferendumvorlage1, expectedReferendumvorlage2));
 
             Assertions.assertThat(result).isEqualTo(expectedResult);
         }
